\name{CorrConf-package}
\alias{CorrConf-package}
\alias{CorrConf}
\docType{package}
\title{
\packageTitle{CorrConf}
}
\description{
\packageDescription{CorrConf}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{CorrConf}
\packageIndices{CorrConf}

}
\author{
\packageAuthor{CorrConf}

Maintainer: \packageMaintainer{CorrConf}
}
\references{

}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\keyword{ Factor analysis }
\keyword{ Correlation }
\seealso{

}
\examples{

}
