% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EstimateCorr_Kg0.R
\name{CreateV}
\alias{CreateV}
\title{Create a covariance matrix V from a list of matrices}
\usage{
CreateV(B, Rho)
}
\arguments{
\item{B}{A list of positive semi-definite matrices}

\item{Rho}{A vector of variance multipliers. length(B) = length(Rho)}
}
\value{
A covariance matrix
}
\description{
Create a covariance matrix V from a list of matrices
}
