% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EstimateCorr_Kg0.R
\name{sqrt.mat}
\alias{sqrt.mat}
\title{Compute the square-root of a symmetric, positive semi-definite matrix}
\usage{
\method{sqrt}{mat}(X)
}
\arguments{
\item{X}{An \code{n} x \code{n} symmetric psd matrix}
}
\value{
A matrix V such that V\%*\%V = X
}
\description{
Compute the square-root of a symmetric, positive semi-definite matrix
}
