\name{CorrConf-package}
\alias{CorrConf-package}
\alias{CorrConf}
\docType{package}
\title{
\packageTitle{CorrConf}
}
\description{
\packageDescription{CorrConf}
}
\details{
The primary functions are ChooseK and EstimateC_complete
}
\author{
\packageAuthor{CorrConf}

Maintainer: \packageMaintainer{CorrConf}
}
\references{
A paper is coming...
}

\keyword{ package }
\keyword{ Factor analysis }
\keyword{ Correlation }
\seealso{

}
\examples{

}