\name{CorrConf-package}
\alias{CorrConf-package}
\alias{CorrConf}
\docType{package}
\title{
\packageTitle{CorrConf}
}
\description{
\packageDescription{CorrConf}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{CorrConf}
\packageIndices{CorrConf}
This package first estimates the number of latent factors, K, using ChooseK. The user should then use this, or some K around the estimated K, to then estimate the latent factors, C, with EEstimateC_complete.
}
\author{
\packageAuthor{CorrConf}

Maintainer: \packageMaintainer{CorrConf}
}
\references{
A paper is on it's way...
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\keyword{ Factor analysis }
\keyword{ Correlation }
\seealso{

}
\examples{

}
