\name{CorrConf-package}
\alias{CorrConf-package}
\alias{CorrConf}
\docType{package}
\title{
\packageTitle{CorrConf}
}
\description{
\packageDescription{CorrConf}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{CorrConf}
\packageIndices{CorrConf}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{CorrConf}

Maintainer: \packageMaintainer{CorrConf}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\keyword{ Factor analysis }
\keyword{ Correlation }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
