% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EstimateC_complete.R
\name{EstimateC_complete}
\alias{EstimateC_complete}
\title{Estimate \code{C} for a given value of \code{K}}
\usage{
EstimateC_complete(
  Y,
  K,
  X = NULL,
  Z = NULL,
  B = NULL,
  A.ine = NULL,
  c.ine = NULL,
  A.equ = NULL,
  Var.0 = NULL,
  Cperp = NULL,
  rho = NULL,
  return.all = T,
  EstVariances = F,
  simpleDelta = F,
  Use.Naive.Omega = F,
  tol.rho = 0.001,
  max.iter.rho = 15,
  return.Bhat = F,
  max.iter.rho.Bhat = 1,
  svd.method = "fast"
)
}
\arguments{
\item{Y}{A \code{p} x \code{n} (number genes/methylation sites x sample size) matrix of observed expression/methylation}

\item{K}{The dimension of the column space of C (i.e. C is a n x K matrix). This should be estimated with ChooseK.}

\item{X}{A \code{n} x \code{d} design matrix of covariates of interest (i.e. disease status)}

\item{Z}{A \code{n} x \code{r} design matrix of nuisance covariates (i.e. known technical factors and the intercept). The default is NONE.}

\item{B}{A list of \code{n} x \code{n} positive semidefinite matrices that determine the correlation structure between samples. For example, these can be kinship matrices, partition matrices that group samples by blocks, diagonal matrices with 0's or 1's along the diagonal grouping samples that have the same residual variance, etc. The software automatically includes the identity if it is not in the span of the user-provided matrices. The default is to treat all samples as independent with the same residual variance.}

\item{A.ine}{A #inequality constaints x b matrix, where b = length(B) = num. variance components. \code{A.ine \%*\% tau >= 0}, where tau = (tau_1^2,...,tau_b^2) are the variance multipliers. The default is all variance multipliers are greater than 0.}

\item{c.ine}{A #inequality constraints vector, where \code{A.ine \%*\% tau - c.ine >= 0}. It is highly recommended users do NOT input anything other than \code{0}. Defaults to \code{0}.}

\item{A.equ}{A #equality constraints x b matrix, where \code{A.equ \%*\% tau = 0}. Defaults to no equality constraints.}

\item{Var.0}{A length \code{b} starting point for tau. If A.equ is specified, this MUST be specified to ensure subsequent iterations lie in the feasible region. Otherwise, the user need not specify a starting point.}

\item{Cperp}{Should be left as NULL. Default is NULL}

\item{rho}{Should be left as NULL. Default is NULL}

\item{return.all}{Return variance multipliers and C for all k = 0,...,K. Default is \code{TRUE}}

\item{EstVariances}{If \code{TRUE} and \code{B} is a single matrix, the variance multipliers for the identity and B are estimated for each unit g=1,...,p. Default if \code{FALSE}}

\item{simpleDelta}{If \code{TRUE}, the model for the residuals is MN(0, delta^2 I_p, V), where V = tau_1^2 B[[1]] + ... + tau_b^2 B[[b]]. If \code{FALSE}, Delta = diag(delta_1^2,...,delta_p^2). Defaults to \code{FALSE}.}

\item{tol.rho}{ICaSE (i.e. sequential PCA) terminates when || tau_j/||tau_j||_2 - tau_\{j-1\}/||tau_\{j-1\}||_2 ||_2 <= tol.rho. Default is \code{1e-3}.}

\item{max.iter.rho}{Maximum number of ICaSE iterations. Default is \code{15}.}

\item{return.Bhat}{If \code{TRUE}, the effects of interest are estimated using GLS with the model Y ~ MN(BX' + LC', Delta, V), where Delta = diag(delta.1^2,...,delta.p^2) and V are re-estimated my REML using the estimated C. Default is \code{FALSE}}

\item{svd.method}{Vestige of previous versions. Should not be altered by the user.}
}
\value{
A list \item{X}{The \code{n} x \code{d} model matrix} \item{Z}{The \code{n} x \code{r} model matrix of nuisance covariates, if specified} \item{C}{The estimate for the latent factors to be used in downstream analyses. The subsequent design matrix to use is [X Z C].} \item{Omega}{The estimate for the correlation between X and C} \item{Cperp}{The part of C orthogonal to \code{X}. If \code{return.all=TRUE}, it returns \code{Cperp} for all k = 1,...,K} \item{rho}{A \code{K+1} x #variance-components matrix, giving the variance multipliers for each k = 0,...,K. rho and tau are used interchangeably in this software.} \item{Delta.hat}{A p-vector of estimates for Delta, when we assume Y ~ MN_\{p x n\}(Beta X' + G Z' + L C', Delta, V) where Delta = diag(delta_1^2,...,delta_p^2). Returned only if \code{return.Bhat=TRUE}} \item{Bhat}{GLS estimate for the regression coefficients of interest, Beta, assuming Y ~ MN_\{p x n\}(Beta X' + G Z' + L C', Delta, V). If \code{B} is specified, this estimate can be improved by using a more accurate model where V is assumed gene/methylation site dependent. Returned only if \code{return.Bhat=TRUE}} \item{pvalues}{P-values for Beta. Returned only if \code{return.Bhat=TRUE}.} \item{tscores}{The \code{p} x \code{d} matrix of t-scores} \item{df.tscores}{The degrees of freedom for the t-scores} \item{tau.Bhat}{The tau estimated by REML under the model Y ~ MN_\{p x n\}(Beta X' + G Z' + L C', Delta, V(tau)). Returned only if \code{return.Bhat=TRUE}}
}
\description{
Estimate \code{C} for a given value of \code{K}. The user can also estimate Beta, the effect due to the covariates of interest by specifying \code{return.Bhat=T}.
}
